+++ school21/42cursus/libft/ft_memcpy.c
@@ -12,7 +12,7 @@
 
 #include "libft.h"
 
-void	*ft_memcpy(void *dst, const void *src, int len)
+void	*ft_memcpy(void *dst, const void *src, size_t len)
 {
 	unsigned char		*ptr_d;
 	unsigned char		*ptr_s;
